+++ eslint.config.js
@@ -1,0 +1,3 @@
+import baseConfig from "@blaze-cardano/eslint-config";
+
+export default baseConfig;