+++ packages/blaze-data/jest.config.ts
@@ -1,2 +1,0 @@
-import config from "@blaze-cardano/jest-config/base.jest.config";
-export default config;