+++ packages/blaze-jest-config/base.jest.config.ts
@@ -1,7 +1,0 @@
-import type { JestConfigWithTsJest } from "ts-jest";
-
-const config: JestConfigWithTsJest = {
-  preset: "ts-jest",
-};
-
-export default config;